refactor: move "send_file" to widget_utils
